Update types for JS SDK
